* Export as json dump for mongodb * Fixed uint64 values ordering

+++ QDbcTool/Alphanum.cpp
@@ -12,12 +12,12 @@
  *    \param    l - строка слева.
  *    \param    r - строка справа.
  *
- *    \return 
+ *    \return
  *        l<r - отрицательное значение;
  *        l>r - положительное значение;
  *        l=r - нуль.
  */
-int compare(QString l, QString r)
+qint64 compare(QString l, QString r)
 {
     enum Mode{STRING, NUMBER} mode = STRING;
     int size;
@@ -27,7 +27,7 @@ int compare(QString l, QString r)
         size = r.size();
 
     int i = 0;
-    
+
     // бежим по обоим строчкам в право до позиции "size-1"
     while(i < size){
         if (mode == STRING){
@@ -55,7 +55,7 @@ int compare(QString l, QString r)
             unsigned long long lnum = 0, rnum = 0;
             int li = i, ri = i; // локальные индексы
             int ld = 0, rd = 0; // цифры
-            
+
             // собираем левое число
             while(li < l.size()){
                 ld = l.at(li).digitValue();
@@ -63,7 +63,7 @@ int compare(QString l, QString r)
                 lnum = lnum*10 + ld;
                 li++;
             }
-            
+
             // собираем правое число
             while(ri < r.size()){
                 rd = r.at(ri).digitValue();
@@ -71,10 +71,10 @@ int compare(QString l, QString r)
                 rnum = rnum*10 + rd;
                 ri++;
             }
-            
+
             long long delta = lnum - rnum;
             if (delta) return delta;
-            
+
             // числа одинаковы
             mode = STRING;
             if (li <= ri)
@@ -86,7 +86,7 @@ int compare(QString l, QString r)
     // Сюда попадём если обе строки до позиции "size-1" одинаковы
     if (i < r.size()) return -1;
     if (i < l.size()) return +1;
-    
+
     // Обе сроки полностью одинаковы
-    return 0;    
+    return 0;
 }
